+++ psycopg/psycopg/transaction.py
@@ -7,6 +7,7 @@ Transaction context managers returned by Connection.transaction()
 from __future__ import annotations
 
 import logging
+from enum import Enum
 from types import TracebackType
 from typing import TYPE_CHECKING, Any, Generic
 from collections.abc import Iterator
@@ -51,6 +52,21 @@ class OutOfOrderTransactionNesting(e.ProgrammingError):
 
 
 class BaseTransaction(Generic[ConnectionType]):
+    class Status(str, Enum):
+        """
+        The current status of a transaction.
+
+        Exposed as the `~psycopg.Transaction.status` attribute of
+        `~psycopg.Transaction` and `~psycopg.AsyncTransaction` objects.
+        """
+
+        NOT_STARTED = "not_started"
+        ACTIVE = "active"
+        COMMITTED = "committed"
+        FAILED = "failed"
+        ROLLED_BACK_EXPLICITLY = "rolled_back_explicitly"
+        ROLLED_BACK_WITH_ERROR = "rolled_back_with_error"
+
     def __init__(
         self,
         connection: ConnectionType,
@@ -61,7 +77,7 @@ class BaseTransaction(Generic[ConnectionType]):
         self.pgconn = self._conn.pgconn
         self._savepoint_name = savepoint_name or ""
         self.force_rollback = force_rollback
-        self._entered = self._exited = False
+        self.status = self.Status.NOT_STARTED
         self._outer_transaction = False
         self._stack_index = -1
 
@@ -77,20 +93,13 @@ class BaseTransaction(Generic[ConnectionType]):
     def __repr__(self) -> str:
         cls = f"{self.__class__.__module__}.{self.__class__.__qualname__}"
         info = connection_summary(self.pgconn)
-        if not self._entered:
-            status = "inactive"
-        elif not self._exited:
-            status = "active"
-        else:
-            status = "terminated"
-
         sp = f"{self.savepoint_name!r} " if self.savepoint_name else ""
-        return f"<{cls} {sp}({status}) {info} at 0x{id(self):x}>"
+        return f"<{cls} {sp}({self.status.value}) {info} at 0x{id(self):x}>"
 
     def _enter_gen(self) -> PQGen[None]:
-        if self._entered:
+        if self.status != self.Status.NOT_STARTED:
             raise TypeError("transaction blocks can be used only once")
-        self._entered = True
+        self.status = self.Status.ACTIVE
 
         self._push_savepoint()
         for command in self._get_enter_commands():
@@ -123,7 +132,7 @@ class BaseTransaction(Generic[ConnectionType]):
 
     def _commit_gen(self) -> PQGen[None]:
         ex = self._pop_savepoint("commit")
-        self._exited = True
+        self.status = self.Status.COMMITTED
         if ex:
             raise ex
 
@@ -135,7 +144,12 @@ class BaseTransaction(Generic[ConnectionType]):
             logger.debug(f"{self._conn}: Explicit rollback from: ", exc_info=True)
 
         ex = self._pop_savepoint("rollback")
-        self._exited = True
+
+        if isinstance(exc_val, Rollback) or self.force_rollback:
+            self.status = self.Status.ROLLED_BACK_EXPLICITLY
+        else:
+            self.status = self.Status.ROLLED_BACK_WITH_ERROR
+
         if ex:
             raise ex
 
@@ -235,7 +249,10 @@ class Transaction(BaseTransaction["Connection[Any]"]):
 
     @property
     def connection(self) -> Connection[Any]:
-        """The connection the object is managing."""
+        """The connection the object is managing.
+
+        :type: `Connection`
+        """
         return self._conn
 
     def __enter__(self) -> Self:
@@ -253,6 +270,7 @@ class Transaction(BaseTransaction["Connection[Any]"]):
             with self._conn.lock:
                 return self._conn.wait(self._exit_gen(exc_type, exc_val, exc_tb))
         else:
+            self.status = self.Status.FAILED
             return False
 
 
@@ -265,6 +283,10 @@ class AsyncTransaction(BaseTransaction["AsyncConnection[Any]"]):
 
     @property
     def connection(self) -> AsyncConnection[Any]:
+        """The connection the object is managing.
+
+        :type: `AsyncConnection`
+        """
         return self._conn
 
     async def __aenter__(self) -> Self:
@@ -282,4 +304,5 @@ class AsyncTransaction(BaseTransaction["AsyncConnection[Any]"]):
             async with self._conn.lock:
                 return await self._conn.wait(self._exit_gen(exc_type, exc_val, exc_tb))
         else:
+            self.status = self.Status.FAILED
             return False